% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R, R/square.R
\name{cube}
\alias{cube}
\alias{square}
\title{Apply a power}
\usage{
cube(x)

square(x)
}
\arguments{
\item{x}{Vector of numberics}

\item{x}{Vector of numberics}
}
\value{
The vector \code{x}, with the corresponding power applied.

The vector \code{x}, with the corresponding power applied.
}
\description{
Apply a power

Apply a power
}
