% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powers2-package.R
\docType{package}
\name{powers2-package}
\alias{powers2}
\alias{powers2-package}
\title{powers2: Apply powers to your vectors}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\details{
Details go here
}
\author{
\strong{Maintainer}: 

Other contributors:
\itemize{
  \item Steve chignell \email{steve.chignell@protonmail.com}
}

}
\keyword{internal}
